Implement common::str_util::ToBool
This function does case-insensitive comparison for strings "true" and
"false", and return nullopt if neither are found.

This can break compatibility with older configs that use "1", "yes", and
"y" as alternatives to true, but the .cfg files tend to replace those
entries with "True". UI.ini is the only case where a backwards
compatability option is added, because it writes "0" and "1" as values
for "Minimized".